+++ src/mesh.cpp
@@ -1,6 +1,6 @@
 #include "mesh.hpp"
 
-bool Mesh::intersect(const Ray &r, Hit &h, float tmin) {
+bool Mesh::intersect(const Ray &r, Hit &h, double tmin) {
 
     // Optional: Change this brute force method into a faster one.
     bool result = false;
