+++ src/utils.cpp
@@ -3,8 +3,8 @@
 #include <iostream>
 
 Vector3f reflect(const Vector3f &I, const Vector3f &N){return I+Vector3f::dot(-I,2.*N)*N;}
-Vector3f refract(const Vector3f &I, const Vector3f &N, float refractivity){
-    float cosd = Vector3f::dot(I,N), sin2d = 1-sqr(cosd), eta, ratio;
+Vector3f refract(const Vector3f &I, const Vector3f &N, double refractivity){
+    double cosd = Vector3f::dot(I,N), sin2d = 1-sqr(cosd), eta, ratio;
     Vector3f Nd = cosd<0?cosd=-cosd,eta=refractivity,-N:(eta=1/refractivity,N);
     return (ratio=sqr(eta)*sin2d)<=1.?I*eta+Nd*(eta*cosd-sqrt(1-ratio)):Vector3f::ZERO;
 }
@@ -21,6 +21,6 @@ void WriteByte( FILE* file, unsigned char b ) {
     assert( success == 1 );
 }
 
-unsigned char ClampColorComponent( float c ) {
+unsigned char ClampColorComponent( double c ) {
     int tmp = int( c * 255 ); return ( unsigned char ) (tmp>=0 ? tmp<=255 ? tmp : 255 : 0);
 }
